CLOOG:  Use arm-eabi-4.9 as a templete for all other scripts (2/2)
Big thanks to @Cl3Kener for pointing out a bug with the toolchain configure
being out of place causing the reset of the script to be null.  This is also now
fixed by using "" around the config option that is causing the issues by using
bash commands in --enable-eh-frame-hdr-for-static

Cloog will now build isl and osl as bundled, so this also emlimates the need to
build isl seperately which is now respectively removed.

Signed-off-by: Paul Beeler <pbeeler80@gmail.com>